requirements.txt
accelerate
aiohttp
aiosignal
annotated-types
ansicon
antlr4-python3-runtime
anyio
appdirs
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
async-timeout
attrs
Babel
backcall
backoff
beautifulsoup4
bitsandbytes
black
bleach
blessed
blis
catalogue
cbor2
certifi
cffi
chardet
charset-normalizer
click
cloudpathlib
cloudpickle
code_bert_score
colorama
coloredlogs
comm
confection
contourpy
coolname
croniter
cryptography
curl_cffi
cycler
cymem
dataclasses-json
datasets
dateutils
debugpy
decorator
deepdiff
defusedxml
Deprecated
dill
distro
docker-pycreds
duckduckgo_search
evaluate
exceptiongroup
executing
faiss-cpu
Farama-Notifications
fastapi
fastjsonschema
filelock
fonttools
fqdn
frozenlist
fsspec
gitdb
GitPython
greenlet
gymnasium
h11
httpcore
httpx
huggingface-hub
humanfriendly
hydra-core
idna
iniconfig
inquirer
ipykernel
ipython
ipython-genutils
ipywidgets
isoduration
isort
itsdangerous
jedi
Jinja2
jinxed
joblib
json5
jsonlines
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchainhub
langcodes
langdetect
langgraph
langsmith
lizard
lxml
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
matplotlib-venn
mdurl
mistune
mpmath
multidict
multiprocess
murmurhash
mypy
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
omegaconf
openai
optimum
ordered-set
orjson
outcome
overrides
pandas
pandocfilters
parso
pathspec
pathtools
pathy
pickleshare
Pillow
platformdirs
plotly
pluggy
portalocker
preshed
prometheus_client
prompt-toolkit
protobuf
psutil
pure-eval
pyarrow
pyarrow-hotfix
pycparser
pydantic
pydantic_core
PyDriller
PyGithub
Pygments
PyJWT
PyNaCl
pyparsing
pyreadline3
PySocks
pytest
python-dateutil
python-dotenv
python-editor
python-json-logger
python-multipart
pytorch-lightning
pytz
PyYAML
pyzmq
qtconsole
QtPy
rank-bm25
readchar
referencing
regex
requests
responses
rich
rpds-py
ruamel.yaml
ruamel.yaml.clib
ruff
sacrebleu
safetensors
scikit-learn
scipy
seaborn
selenium
self
Send2Trash
sentencepiece
sentry-sdk
setproctitle
setuptools-scm
six
smart-open
smmap
sniffio
sortedcontainers
soupsieve
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
stack-data
starlette
starsessions
sympy
tabulate
tenacity
terminado
thinc
threadpoolctl
tiktoken
tinycss2
tokenize-rt
tokenizers
tomli
torch
torchmetrics
tornado
tqdm
traitlets
transformers
trio
trio-websocket
typer
types-python-dateutil
types-pytz
types-PyYAML
types-requests
typing-inspect
typing_extensions
tzdata
uri-template
urllib3
uvicorn
wandb
wasabi
wcwidth
weasel
webcolors
webdriver-manager
webencodings
websocket-client
websockets
widgetsnbextension
wrapt
wsproto
XlsxWriter
xxhash
yarl